requirements.txt
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
psutil==7.0.0
pyaes==1.6.1
pyasn1==0.6.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
pytweening==1.2.0
pywin32==311; sys_platform == "win32"
requests==2.32.4
rsa==4.9.1
six==1.17.0
Telethon==1.40.0
urllib3==2.5.0
Werkzeug==3.1.3
WMI==1.5.1
yt-dlp==2025.6.30
